\name{print.bigWig}
\alias{print.bigWig}
\alias{print}

\title{Print Information on bigWig Files}
\description{Print information on loaded UCSC bigWig files.}
\usage{
\method{print}{bigWig}(x, \dots)
}


\arguments{
\item{x}{a bigWig object}
\item{...}{other further arguments passed to or from other methods.}
}

\details{Displays information bigWig instance in the same format as the command line tool bigWigInfo.}

\author{André Luís Martins}

\seealso{See \code{\link{load.bigWig}} to load a bigWig}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ bigWig }
\keyword{ print }
