\name{bigWig-package}
\alias{bigWig-package}
\alias{bigWig}
\docType{package}
\title{
bigWig Query Library
~~ bigWig ~~
}
\description{R interface to query bigWig files. C implementation relies
  on UCSC Genome Browser code.
}
\details{
\tabular{ll}{
Package: \tab bigWig\cr
Type: \tab Package\cr
Version: \tab 0.1-14\cr
Date: \tab 2013-05-31\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}
bigWig files should be loaded with the \code{load.bigWig} function and
then queries can be conducted with \code{query.bigWig}. There is support
for loading both local and remote files (http,ftp). Some utility
functions are also provided to build meta plots.
}
\author{
André Luís Martins

Maintainer: André Luís Martins <alm253@cornell.edu>
}
\references{}
\keyword{ package }
\seealso{
}
\examples{
~~ simple examples of the most important functions ~~
}
